fix styling and save attractions selected by day to DB

+++ client/src/components/PackageDays.jsx
@@ -3,12 +3,14 @@ import axios from 'axios';
 import DatePicker from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
 import { useParams } from 'react-router-dom';
-import { TextField, Button, Typography, Modal, IconButton, Card } from '@mui/material';
+import { TextField, Button, Card, Typography, Modal, IconButton } from '@mui/material';
 import CalendarTodayIcon from '@mui/icons-material/CalendarToday';
 import DeleteIcon from '@mui/icons-material/Delete';
 import CloseIcon from '@mui/icons-material/Close';
 import AddIcon from '@mui/icons-material/Add';
 import RemoveIcon from '@mui/icons-material/Remove';
+import TodayIcon from '@mui/icons-material/Today';
+import { Fab } from '@mui/material';
 import ActivitiesListItem from './ActivitiesListItem';
 
 const PackageDays = () => {
@@ -24,6 +26,10 @@ const PackageDays = () => {
   const [selectedAttractions, setSelectedAttractions] = useState([]);
   const [fetchAttractions, setFetchAttractions] = useState(false);
   const [activeDayId, setActiveDayId] = useState(null);
+//fetch all attractions that belong to a specific day that was clicked and populate the array that is conditional to show the - sign"
+
+
+
 
 
   useEffect(() => {
@@ -60,7 +66,7 @@ const PackageDays = () => {
         title: newDayTitle,
         description: newDayDescription,
         date: newDayDate,
-        attractions: selectedAttractions,
+        attractions: selectedAttractions, // Pass the selected attractions to the backend
       };
       await axios.post(`/addDay/${packageId}`, newDay);
       setDayAdded(true);
@@ -77,15 +83,19 @@ const PackageDays = () => {
   };
 
   const handleDeleteDay = async (dayId) => {
-    try {
-      await axios.delete(`/deleteDay/${dayId}`);
-      setDayAdded(true);
-    } catch (error) {
-      console.error(error);
+    const confirmDelete = window.confirm('Are you sure you want to delete this day?');
+    if (confirmDelete) {
+      try {
+        await axios.delete(`/deleteDay/${dayId}`);
+        setDayAdded(true);
+      } catch (error) {
+        console.error(error);
+      }
     }
   };
 
   const handleOpenModal = (dayid) => {
+    setSelectedAttractions([]);
     setFetchAttractions(true);
     setIsOpen(true);
     console.log(dayid)
@@ -98,6 +108,7 @@ const PackageDays = () => {
 
   const handleAddAttraction = (attractionId) => {
     setSelectedAttractions((prevSelectedAttractions) => [...prevSelectedAttractions, attractionId]);
+  
   };
 
   const handleRemoveAttraction = (attractionId) => {
@@ -106,33 +117,67 @@ const PackageDays = () => {
     );
   };
 
-  const handleSaveAttractions = () => {
-    console.log(selectedAttractions,activeDayId);
+  const handleSaveAttractions = async () => {
+    console.log(selectedAttractions, activeDayId);
     handleCloseModal();
+  
+    try {
+      await axios.post(`/insertAttractionsByDay/${activeDayId}`, { attractions: selectedAttractions });
+      setSelectedAttractions([]);
+      console.log('successfully apdated attraction list for this day')
+    } catch (error) {
+      console.error(error);
+    }
   };
 
   const renderDays = () => {
     return days
       .sort((a, b) => a.date.localeCompare(b.date))
       .map((day, index) => (
-        <Card key={day.day_id} style={{ width: 300, margin: 12 }}>
-          <div style={{ padding: '16px' }}>
-            <Typography variant="h6" style={{ fontStyle: 'italic', textDecoration: 'underline' }}>Day {index + 1} </Typography>
-            <Typography variant="h6" >{day.day_title}</Typography>
+        <Card
+          key={day.day_id}
+          style={{
+            marginBottom: '10px',
+            padding: '10px',
+            display: 'flex',
+            justifyContent: 'space-between',
+            alignItems: 'center',
+          }}
+        >
+          <div style={{ display: 'flex', flexDirection: 'column' }}>
+            <Typography variant="h5">
+              Day {index + 1}{' '}
+              <Typography variant="h6" component="span">
+                {day.date.substring(0, 10)} <TodayIcon />
+              </Typography>
+            </Typography>
+            <Typography variant="h6"> {day.day_title}</Typography>
             <Typography>{day.day_description}</Typography>
-            <Typography>Date: {day.date.substring(0, 10)}</Typography>
           </div>
-          <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', padding: '8px' }}>
-            <Button variant="contained" color="primary" style={{ margin: 8, background: "#51D4BF", borderRadius: 20, maxHeight: 30 }} onClick={()=>handleOpenModal(day.day_id)}>
-              Add 
-            </Button>
-            <IconButton onClick={() => handleDeleteDay(day.day_id)} style={{ color: 'grey' }}>
-              <DeleteIcon />
-            </IconButton>
+          <div style={{ display: 'flex', alignItems: 'center' }}>
+            <Fab
+              color="primary"
+              aria-label="Add Attraction"
+              onClick={()=>handleOpenModal(day.day_id)}
+              style={{
+                backgroundColor: '#51D4BF',
+                width: '56px',
+                height: '56px',
+              }}
+            >
+              <AddIcon style={{ fontSize: '32px' }} />
+            </Fab>
+            <DeleteIcon
+              style={{ color: 'grey', cursor: 'pointer', marginLeft: '10px', fontSize: '40px' }}
+              onClick={() => handleDeleteDay(day.day_id)}
+            />
           </div>
         </Card>
       ));
-  };
+  }
+
+
+
   return (
     <>
       <form onSubmit={handleAddDay} style={{ marginTop: '30px' }}>
@@ -174,12 +219,28 @@ const PackageDays = () => {
               </div>
             )}
           </div>
-          <Button type="submit" variant="contained" color="primary" style={{ marginLeft: '20px',background: "#51D4BF", borderRadius: 20 }}>Add</Button>
+          <Button type="submit" variant="contained" color="primary" style={{ marginLeft: '20px', background: '#51D4BF', borderRadius: 20, height: '50px' }}>
+            Add
+          </Button>
         </div>
       </form>
       <div style={{ margin: '20px' }}>{renderDays()}</div>
       <Modal open={isOpen} onClose={handleCloseModal}>
-        <div style={{ position: 'absolute', top: '50%', left: '50%', transform: 'translate(-50%, -50%)', backgroundColor: 'white', padding: '16px', borderRadius: '8px', maxHeight: '90vh', overflowY: 'auto' }}>
+        <Card
+          style={{
+            position: 'absolute',
+            top: '50%',
+            left: '50%',
+            transform: 'translate(-50%, -50%)',
+            width: '80%',
+            backgroundColor: 'white',
+            padding: '16px',
+            borderRadius: '8px',
+            maxHeight: '90vh',
+            overflowY: 'auto',
+
+          }}
+        >
           <div style={{ display: 'flex', justifyContent: 'flex-end' }}>
             <IconButton aria-label="close" onClick={handleCloseModal}>
               <CloseIcon />
@@ -188,35 +249,81 @@ const PackageDays = () => {
           <Typography variant="h5">Attractions</Typography>
           <div style={{ display: 'flex', flexWrap: 'wrap', justifyContent: 'space-between' }}>
             {attractions.map((attraction) => (
-              <div key={attraction.id} style={{ marginBottom: '20px' }}>
-                <img src={attraction.pictures[0]} alt="Attraction" style={{ width: '100%', height: '200px', objectFit: 'cover', borderRadius: '8px' }} />
-                <div style={{ display: 'flex', flexWrap: 'wrap', justifyContent: 'space-between', paddingRight: '16px' }}>
-                  <div>
-                    <Typography variant="h6">{attraction.country}</Typography>
-                    <Typography variant="subtitle1">{attraction.city}</Typography>
+              <Card
+                key={attraction.id}
+                style={{
+                  width: '30%',
+                  marginBottom: '20px',
+                  display: 'flex',
+                  flexDirection: 'column',
+                  alignItems: 'center',
+                  padding: '16px',
+                  borderRadius: '8px',
+                  boxShadow: '0px 2px 10px rgba(0, 0, 0, 0.1)',
+                }}
+              >
+                <div style={{ maxHeight: '300px', overflow: 'auto' }}>
+                  <img
+                    src={attraction.pictures[0]}
+                    alt="Attraction"
+                    style={{ width: '100%', height: '200px', objectFit: 'cover', borderRadius: '8px' }}
+                  />
+                  <div style={{ display: 'flex', flexWrap: 'wrap', justifyContent: 'space-between', paddingRight: '16px' }}>
+                    <div>
+                      <Typography variant="h6">{attraction.country}</Typography>
+                      <Typography variant="subtitle1">{attraction.city}</Typography>
+                    </div>
+                    <div>
+                      <Typography variant="h6">Price: ${attraction.price}</Typography>
+                      <Typography>Duration: {attraction.duration / 60}hours</Typography>
+                    </div>
                   </div>
-                  <div>
-                  <Typography variant="h6">Price: ${attraction.price}</Typography>
-                  <Typography>Duration: {attraction.duration/60}hours</Typography>
-                  </div>
+                  <Typography>{attraction.description}</Typography>
+
                 </div>
-                <Typography>{attraction.description}</Typography>
-                <IconButton aria-label="Add attraction" onClick={() => handleAddAttraction(attraction.attraction_id)}>
+                <IconButton
+                  aria-label="Add attraction"
+                  onClick={() => handleAddAttraction(attraction.attraction_id)}
+                  style={{ marginTop: 'auto' }}
+                >
                   <AddIcon />
                 </IconButton>
                 {selectedAttractions.includes(attraction.attraction_id) && (
-                  <IconButton aria-label="Remove attraction" onClick={() => handleRemoveAttraction(attraction.attraction_id)}>
+                  <IconButton
+                    aria-label="Remove attraction"
+                    onClick={() => handleRemoveAttraction(attraction.attraction_id)}
+                  >
                     <RemoveIcon />
                   </IconButton>
                 )}
-              </div>
+              </Card>
             ))}
           </div>
-          <Button variant="contained" color="primary" onClick={handleSaveAttractions}>Save</Button>
-        </div>
+          <Button
+            variant="contained"
+            color="primary"
+            onClick={handleSaveAttractions}
+            style={{
+              position: 'sticky',
+              bottom: '10px',
+              left: '90vh',
+              borderRadius: '10%',
+              width: '100px',
+              height: '50px',
+              zIndex: '999',
+              background: '#51D4BF',
+            }} >
+            Save
+          </Button>
+        </Card>
       </Modal>
+
+
+
     </>
   );
 };
 
 export default PackageDays;
+
+
